Use proxy for agent homepage negotiation
